Separate AppBackground into its own file
