+++ frontend/app/app.tsx
@@ -1,35 +1,22 @@
 // Copyright 2024, Command Line Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-import { useWaveObjectValue } from "@/app/store/wos";
 import { Workspace } from "@/app/workspace/workspace";
 import { ContextMenuModel } from "@/store/contextmenu";
-import {
-    PLATFORM,
-    WOS,
-    atoms,
-    createBlock,
-    getApi,
-    globalStore,
-    removeFlashError,
-    useSettingsPrefixAtom,
-} from "@/store/global";
+import { PLATFORM, atoms, createBlock, globalStore, removeFlashError, useSettingsPrefixAtom } from "@/store/global";
 import { appHandleKeyDown } from "@/store/keymodel";
-import { getWebServerEndpoint } from "@/util/endpoints";
 import { getElemAsStr } from "@/util/focusutil";
 import * as keyutil from "@/util/keyutil";
 import * as util from "@/util/util";
-import useResizeObserver from "@react-hook/resize-observer";
 import clsx from "clsx";
 import Color from "color";
-import * as csstree from "css-tree";
 import debug from "debug";
-import * as jotai from "jotai";
+import { Provider, useAtomValue } from "jotai";
 import "overlayscrollbars/overlayscrollbars.css";
-import * as React from "react";
+import { Fragment, useEffect, useState } from "react";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
-import { debounce } from "throttle-debounce";
+import { AppBackground } from "./app-bg";
 import "./app.less";
 import { CenteredDiv } from "./element/quickelems";
 
@@ -37,7 +24,6 @@ const dlog = debug("wave:app");
 const focusLog = debug("wave:focus");
 
 const App = () => {
-    let Provider = jotai.Provider;
     return (
         <Provider store={globalStore}>
             <AppInner />
@@ -123,8 +109,8 @@ async function handleContextMenu(e: React.MouseEvent<HTMLDivElement>) {
 
 function AppSettingsUpdater() {
     const windowSettingsAtom = useSettingsPrefixAtom("window");
-    const windowSettings = jotai.useAtomValue(windowSettingsAtom);
-    React.useEffect(() => {
+    const windowSettings = useAtomValue(windowSettingsAtom);
+    useEffect(() => {
         const isTransparentOrBlur =
             (windowSettings?.["window:transparent"] || windowSettings?.["window:blur"]) ?? false;
         const opacity = util.boundNumber(windowSettings?.["window:opacity"] ?? 0.8, 0, 1);
@@ -163,7 +149,7 @@ function AppFocusHandler() {
     return null;
 
     // for debugging
-    React.useEffect(() => {
+    useEffect(() => {
         document.addEventListener("focusin", appFocusIn);
         document.addEventListener("focusout", appFocusOut);
         document.addEventListener("selectionchange", appSelectionChange);
@@ -183,122 +169,8 @@ function AppFocusHandler() {
     return null;
 }
 
-function encodeFileURL(file: string) {
-    const webEndpoint = getWebServerEndpoint();
-    return webEndpoint + `/wave/stream-file?path=${encodeURIComponent(file)}&no404=1`;
-}
-
-function processBackgroundUrls(cssText: string): string {
-    if (util.isBlank(cssText)) {
-        return null;
-    }
-    cssText = cssText.trim();
-    if (cssText.endsWith(";")) {
-        cssText = cssText.slice(0, -1);
-    }
-    const attrRe = /^background(-image)?\s*:\s*/i;
-    cssText = cssText.replace(attrRe, "");
-    const ast = csstree.parse("background: " + cssText, {
-        context: "declaration",
-    });
-    let hasUnsafeUrl = false;
-    csstree.walk(ast, {
-        visit: "Url",
-        enter(node) {
-            const originalUrl = node.value.trim();
-            if (
-                originalUrl.startsWith("http:") ||
-                originalUrl.startsWith("https:") ||
-                originalUrl.startsWith("data:")
-            ) {
-                return;
-            }
-            // allow file:/// urls (if they are absolute)
-            if (originalUrl.startsWith("file://")) {
-                const path = originalUrl.slice(7);
-                if (!path.startsWith("/")) {
-                    console.log(`Invalid background, contains a non-absolute file URL: ${originalUrl}`);
-                    hasUnsafeUrl = true;
-                    return;
-                }
-                const newUrl = encodeFileURL(path);
-                node.value = newUrl;
-                return;
-            }
-            // allow absolute paths
-            if (originalUrl.startsWith("/") || originalUrl.startsWith("~/")) {
-                const newUrl = encodeFileURL(originalUrl);
-                node.value = newUrl;
-                return;
-            }
-            hasUnsafeUrl = true;
-            console.log(`Invalid background, contains an unsafe URL scheme: ${originalUrl}`);
-        },
-    });
-    if (hasUnsafeUrl) {
-        return null;
-    }
-    const rtnStyle = csstree.generate(ast);
-    if (rtnStyle == null) {
-        return null;
-    }
-    return rtnStyle.replace(/^background:\s*/, "");
-}
-
-function AppBackground() {
-    const bgRef = React.useRef<HTMLDivElement>(null);
-    const tabId = jotai.useAtomValue(atoms.activeTabId);
-    const [tabData] = useWaveObjectValue<Tab>(WOS.makeORef("tab", tabId));
-    const bgAttr = tabData?.meta?.bg;
-    const style: React.CSSProperties = {};
-    if (!util.isBlank(bgAttr)) {
-        try {
-            const processedBg = processBackgroundUrls(bgAttr);
-            if (!util.isBlank(processedBg)) {
-                const opacity = util.boundNumber(tabData?.meta?.["bg:opacity"], 0, 1) ?? 0.5;
-                style.opacity = opacity;
-                style.background = processedBg;
-                const blendMode = tabData?.meta?.["bg:blendmode"];
-                if (!util.isBlank(blendMode)) {
-                    style.backgroundBlendMode = blendMode;
-                }
-            }
-        } catch (e) {
-            console.error("error processing background", e);
-        }
-    }
-    const getAvgColor = React.useCallback(
-        debounce(30, () => {
-            if (
-                bgRef.current &&
-                PLATFORM !== "darwin" &&
-                bgRef.current &&
-                "windowControlsOverlay" in window.navigator
-            ) {
-                const titlebarRect: Dimensions = (window.navigator.windowControlsOverlay as any).getTitlebarAreaRect();
-                const bgRect = bgRef.current.getBoundingClientRect();
-                if (titlebarRect && bgRect) {
-                    const windowControlsLeft = titlebarRect.width - titlebarRect.height;
-                    const windowControlsRect: Dimensions = {
-                        top: titlebarRect.top,
-                        left: windowControlsLeft,
-                        height: titlebarRect.height,
-                        width: bgRect.width - bgRect.left - windowControlsLeft,
-                    };
-                    getApi().updateWindowControlsOverlay(windowControlsRect);
-                }
-            }
-        }),
-        [bgRef, style]
-    );
-    React.useLayoutEffect(getAvgColor, [getAvgColor]);
-    useResizeObserver(bgRef, getAvgColor);
-
-    return <div ref={bgRef} className="app-background" style={style} />;
-}
-
 const AppKeyHandlers = () => {
-    React.useEffect(() => {
+    useEffect(() => {
         const staticKeyDownHandler = keyutil.keydownWrapper(appHandleKeyDown);
         document.addEventListener("keydown", staticKeyDownHandler);
 
@@ -310,11 +182,11 @@ const AppKeyHandlers = () => {
 };
 
 const FlashError = () => {
-    const flashErrors = jotai.useAtomValue(atoms.flashErrors);
-    const [hoveredId, setHoveredId] = React.useState<string>(null);
-    const [ticker, setTicker] = React.useState<number>(0);
+    const flashErrors = useAtomValue(atoms.flashErrors);
+    const [hoveredId, setHoveredId] = useState<string>(null);
+    const [ticker, setTicker] = useState<number>(0);
 
-    React.useEffect(() => {
+    useEffect(() => {
         if (flashErrors.length == 0 || hoveredId != null) {
             return;
         }
@@ -351,10 +223,10 @@ const FlashError = () => {
 
     function convertNewlinesToBreaks(text) {
         return text.split("\n").map((part, index) => (
-            <React.Fragment key={index}>
+            <Fragment key={index}>
                 {part}
                 <br />
-            </React.Fragment>
+            </Fragment>
         ));
     }
 
@@ -382,10 +254,10 @@ const FlashError = () => {
 };
 
 const AppInner = () => {
-    const prefersReducedMotion = jotai.useAtomValue(atoms.prefersReducedMotionAtom);
-    const client = jotai.useAtomValue(atoms.client);
-    const windowData = jotai.useAtomValue(atoms.waveWindow);
-    const isFullScreen = jotai.useAtomValue(atoms.isFullScreen);
+    const prefersReducedMotion = useAtomValue(atoms.prefersReducedMotionAtom);
+    const client = useAtomValue(atoms.client);
+    const windowData = useAtomValue(atoms.waveWindow);
+    const isFullScreen = useAtomValue(atoms.isFullScreen);
 
     if (client == null || windowData == null) {
         return (
